Revert "Disable SPA"
This reverts commit a499aeb34475a790f3f41d46d13215abf1336d5b.

There is nothing related to the SPA that was confusing the routing.

I have a page named adilson.xyz, and everywhere a link to it is
rendered, instead of it being './Notes/adilson.xyz.html', it is
rendered as './Notes.adilson.xyz'.

This causes the SPA application to keeping adding a '/Notes' path
segment to the url ad infinitum. The non-SPA version just plain
breaks by rendering the site's home page.

The problem lays with whatever is responsible to cook the right
path reference to that file.

Ah, the adilson.xyz.html file is generated, and one can reach it
by adding .html to the url manually.